refactor(frontend): unify message stores
