+++ frontend/src/lib/state/room/messages.svelte.ts
@@ -1,4 +1,3 @@
 export { MessageListStore } from './messages/MessageListStore.svelte';
-export { RoomMessagesStore } from './messages/RoomMessagesStore.svelte';
-export { ThreadMessagesStore } from './messages/ThreadMessagesStore.svelte';
+export { MessagesStore } from './messages/MessagesStore.svelte';
 export { isRootRoomEvent, isThreadEvent } from './messages/filters';
