+++ frontend/src/routes/chat/[serverId]/[roomId]/RoomEventsPane.svelte
@@ -2,7 +2,7 @@
   import { onDestroy } from 'svelte';
   import { useEvent } from '$lib/hooks';
   import { useConnection } from '$lib/state/server/connection.svelte';
-  import { getComposerContext, RoomMessagesStore, type RoomMember } from '$lib/state/room';
+  import { getComposerContext, MessagesStore, type RoomMember } from '$lib/state/room';
   import { getActiveServer } from '$lib/state/activeServer.svelte';
   import { serverRegistry } from '$lib/state/server/registry.svelte';
   import EventList from './EventList.svelte';
@@ -29,7 +29,7 @@
   const jumpState = composerContext.jumpState;
   const currentUser = $derived(serverRegistry.getStore(getActiveServer()).currentUser);
 
-  const store = new RoomMessagesStore(
+  const store = new MessagesStore(
     connection(),
     () => currentUser.user?.id ?? null
   );
